Converted QThread subclass to QObject subclass.
Apparently it is wrong to subclass QThread. I converted to the correct
way to start the thread.

Introduce a progress event so the progress of the normal computation can be
monitored. The computation object wraps the event in a qt signal to pass it
back to the GUI thread so a ProgressDialog can be used.